Update README.md
The learningFAQs_chatbot is a versatile and intelligent chatbot tailored to address frequently asked questions within the educational domain. It serves as an accessible and efficient resource for students, educators, and academic institutions by providing instant answers to queries related to courses, admissions, scholarships, and more. With a focus on enhancing the learning experience, this chatbot is designed to deliver quick, accurate, and personalized responses, making it easier for users to navigate through educational resources and make informed decisions.